Change book and forbid to delete genre or writer that contains in book.

+++ public/scripts/genres.js
@@ -2,17 +2,33 @@
 	let 
 		db = firebase.firestore(),
 		genresCol = db.collection("genres"),
+		booksCol = db.collection("books"),
 
         loadBar = $("#load-bar");
 
 	function removeGenre(genreId) {
 		loadBar.css("display", "block");
 
-		genresCol.doc(genreId).delete()
-		.then(function() {
-			$("[data-id=" + genreId + "]").remove();
-			loadBar.css("display", "none");
-		});
+		booksCol.where("genresId", "array-contains", genreId).get()
+		.then(function(booksQuery) {
+			if (booksQuery.size == 0) {
+				genresCol.doc(genreId).delete()
+				.then(function() {
+					$("[data-id=" + genreId + "]").remove();
+					loadBar.css("display", "none");
+				});
+			} else {
+				loadBar.css("display", "none");
+
+				let books = [];
+
+				booksQuery.docs.forEach(function(doc) {
+					books.push(doc.data().title);
+				});
+
+				alert("Неможливо видалити, бо цей жанр використовується у книгах: " + books.join(", "));
+			}
+		})
 	}
 	/*
 	genresCol.onSnapshot(snapshot => {
